pyproject.toml
[tool.poetry]
name = "luca-bio"
version = "3.6.9"
description = "L.U.C.A - Bio-inspired AI with 369/370 Quality Standard and Neurodiversity Optimization"
authors = ["Lennart Wuchold <lenny@luca.bio>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/lennartwuchold-LUCA/LUCA-AI_369"
repository = "https://github.com/lennartwuchold-LUCA/LUCA-AI_369"
keywords = ["bioinformatics", "resource-allocation", "optimization", "enzymatic-kinetics", "hill-equation", "neurodiversity", "quality-standard", "info-blocks"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
packages = [
    {include = "luca"}
]

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.0"
scipy = "^1.11.0"
jsonschema = "^4.20.0"
matplotlib = "^3.8.0"
pandas = "^2.1.0"
pillow = "^10.0.0"
click = "^8.1.7"

# Backend dependencies (optional)
fastapi = {version = "^0.104.0", optional = true}
uvicorn = {version = "^0.24.0", extras = ["standard"], optional = true}
sqlalchemy = {version = "^2.0.0", optional = true}
anthropic = {version = "^0.71.0", optional = true}
bcrypt = {version = "^4.1.0", optional = true}
pyjwt = {version = "^2.8.0", optional = true}
python-dotenv = {version = "^1.0.0", optional = true}
pydantic = {version = "^2.5.0", optional = true}
pydantic-settings = {version = "^2.1.0", optional = true}
python-multipart = {version = "^0.0.6", optional = true}
email-validator = {version = "^2.1.0", optional = true}

# Universal Root Kernel dependencies (optional)
pyserial = {version = "^3.5", optional = true}
meshtastic = {version = "^2.2.0", optional = true}
qutip = {version = "^4.7.0", optional = true}
requests = {version = "^2.31.0", optional = true}

[tool.poetry.extras]
backend = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "anthropic",
    "bcrypt",
    "pyjwt",
    "python-dotenv",
    "pydantic",
    "pydantic-settings",
    "python-multipart",
    "email-validator"
]
universal = [
    "pyserial",
    "meshtastic",
    "qutip",
    "requests"
]
all = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "anthropic",
    "bcrypt",
    "pyjwt",
    "python-dotenv",
    "pydantic",
    "pydantic-settings",
    "python-multipart",
    "email-validator",
    "pyserial",
    "meshtastic",
    "qutip",
    "requests"
]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-xdist = "^3.5.0"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.0"

[tool.poetry.scripts]
luca = "luca.allocator:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests", "luca_369_370/tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
pythonpath = ["."]
